push swap done?

+++ push_swap/push_swap/lib1.c
@@ -1,6 +1,6 @@
 #include "push_swap.h"
 
-int		ft_atoi(char *str, int *data)
+int	ft_atoi(char *str, int *data)
 {
 	int			sign;
 	long long	result;
@@ -25,7 +25,7 @@ int		ft_atoi(char *str, int *data)
 	return (OK);
 }
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	char	*ptr;
